+++ src/Common/ses.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Lib;
+namespace App\Common;
 
 class Ses
 {